+++ apps/next-client/src/constants/errors.ts
@@ -1,5 +1,6 @@
 export const ERROR_MESSAGES = {
   LOGIN_FAILED: "이메일이나 비밀번호를 다시 확인해주세요.",
+  LOGIN_REQUIRED: "로그인이 필요한 서비스입니다.",
   UNKNOWN_ERROR: "알 수 없는 오류가 발생했습니다.",
   INVALID_CREDENTIAL: "유효하지 않은 자격 증명입니다.",
   LOGIN_ERROR: "로그인 중 문제가 발생했습니다.",
